+++ app/src/main/java/com/alexcfa/yourweather/data/DomainModels.kt
@@ -1,0 +1,17 @@
+package com.alexcfa.yourweather.data
+
+data class HourlyModel(
+    val time: String?,
+    val temperature: Int?,
+    val windSpeed: Int?,
+    val weatherIcons: List<String>,
+    val weatherDescriptions: List<String>?,
+    val precip: Double?,
+    val pressure: Int?
+)
+
+data class CurrentLocationModel(
+    val request: Request?,
+    val location: Location?,
+    val current: Current?
+)
