+++ app/src/main/java/com/alexcfa/yourweather/data/RemoteResult.kt
@@ -1,3 +1,5 @@
+package com.alexcfa.yourweather.data
+
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
